Move `ConstructionObserver` to test helpers
